requirements.txt
django>=4.1
djangorestframework>=3.13
opencv-python>=4.5.5.62
pillow>=8.4.0
django-cors-headers>=3.10.0
django-sslserver>=0.22
drf-yasg
djangorestframework-simplejwt